Bug fix for ajax server side error. Signed-Off-By: Piyush Vijay <piyushvijay.1997@gmail.com>
